# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  auth.Login2FADTO:
    properties:
      session_token:
        description: 登录时返回的临时会话令牌
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      two_factor_code:
        description: 6位TOTP验证码
        example: "123456"
        type: string
    required:
    - session_token
    - two_factor_code
    type: object
  auth.LoginDTO:
    properties:
      account:
        description: 手机号/用户名/邮箱
        example: admin
        type: string
      captcha:
        description: 验证码
        example: "9999"
        type: string
      captcha_id:
        description: 验证码ID
        example: dev-123456
        type: string
      password:
        description: 密码
        example: admin123
        type: string
    required:
    - account
    - captcha
    - captcha_id
    - password
    type: object
  auth.LoginResponseDTO:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      requires_2fa:
        description: 2FA 相关（当需要 2FA 时返回）
        type: boolean
      session_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/auth.UserBriefDTO'
    type: object
  auth.LoginRoleDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  auth.RefreshTokenDTO:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterDTO:
    properties:
      email:
        example: john@example.com
        type: string
      nickname:
        example: Johnny
        maxLength: 50
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
      real_name:
        example: John Doe
        maxLength: 100
        type: string
      signature:
        example: Hello World
        maxLength: 255
        type: string
      username:
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  auth.RegisterResultDTO:
    properties:
      access_token:
        type: string
      email:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  auth.TokenDTO:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  auth.UserBriefDTO:
    properties:
      roles:
        items:
          $ref: '#/definitions/auth.LoginRoleDTO'
        type: array
      user_id:
        type: integer
      username:
        type: string
    type: object
  captcha.GenerateResultDTO:
    properties:
      code:
        description: Code 验证码值（仅开发模式返回）
        type: string
      expire_at:
        description: ExpireAt 过期时间戳（秒）
        type: integer
      id:
        description: ID 验证码ID
        type: string
      image:
        description: Image Base64编码的图片
        type: string
    type: object
...
  response.DataResponse-auth_LoginResponseDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/auth.LoginResponseDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-auth_RegisterResultDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/auth.RegisterResultDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-auth_TokenDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/auth.TokenDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-captcha_GenerateResultDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/captcha.GenerateResultDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/lwmacct/260103-ddd-iam-bc
  description: 基于 DDD + CQRS 架构的可复用模块库
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go DDD Package Library API
  version: "1.0"
paths:
...
  /api/auth/captcha:
    get:
      consumes:
      - application/json
      description: 生成图形验证码用于登录。支持开发模式（通过code和secret参数指定验证码值）
      parameters:
      - description: 开发模式：指定验证码值
        in: query
        name: code
        type: string
      - description: 开发模式：密钥
        in: query
        name: secret
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 验证码生成成功
          schema:
            $ref: '#/definitions/response.DataResponse-captcha_GenerateResultDTO'
        "500":
          description: 生成失败
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 获取验证码
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 使用手机号/用户名/邮箱和密码登录系统，需要提供图形验证码。如果启用了2FA，返回session_token用于后续2FA验证
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功或需要2FA验证
          schema:
            $ref: '#/definitions/response.DataResponse-auth_LoginResponseDTO'
        "401":
          description: 登录失败：凭证无效、验证码错误或账户被禁用
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 登录
      tags:
      - auth
  /api/auth/login/2fa:
    post:
      consumes:
      - application/json
      description: 使用session_token和2FA验证码完成登录（适用于启用了2FA的账户）
      parameters:
      - description: 二次认证凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.Login2FADTO'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/response.DataResponse-auth_TokenDTO'
        "401":
          description: 验证失败：session_token无效或2FA验证码错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 2FA 登录
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用refresh_token获取新的access_token和refresh_token，延长会话有效期
      parameters:
      - description: 刷新令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 令牌刷新成功
          schema:
            $ref: '#/definitions/response.DataResponse-auth_LoginResponseDTO'
        "401":
          description: 刷新令牌无效或已过期
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 刷新令牌
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号，注册成功后自动登录并返回访问令牌
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/response.DataResponse-auth_RegisterResultDTO'
        "400":
          description: 参数错误或用户名/邮箱已存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 注册
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"